requirements.txt
asgiref==3.5.2
Django~=4.1.0
sqlparse==0.4.2
Pillow==9.2.0
flower==1.2.0
celery==5.2.7
stripe==4.0.2
WeasyPrint==56.1
redis==4.3.4
social-auth-app-django==5.4.2
django-bootstrap-v5==1.0.11
django-extensions==3.2.3
weasyprint==56.1
psycopg2-binary==2.9.5